api/requirements.txt
fastapi
pydantic
uvicorn
pymongo
bcrypt
python-dotenv